update testing routine based on cosine similarity and operating responsibility move to operation file

+++ src/matrix/matrix_util.cpp
@@ -33,7 +33,7 @@ double MatrixUtil::CosineSimilarity(const Matrix& lhs, const Matrix& rhs) {
         rhs_ = rhs.Transpose();
     }
 
-    double result = lhs_ * rhs_;
+    double result = static_cast<double>(lhs_ * rhs_);
     result /= Matrix::Norm2(lhs_) * Matrix::Norm2(rhs_);
     return result;
 }
